requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tabulate>=0.9.0
pylint>=3.0.0
